requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
proxy-tools==0.1.0
pycparser==2.21
pythonnet==2.5.2
pywebview==3.6.3
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
